+++ src/Catesta.build.ps1
@@ -254,7 +254,7 @@ Add-BuildTask Test {
         New-Item -Path $testOutPutPath -ItemType Directory | Out-Null
     }
     if (Test-Path -Path $script:UnitTestsPath) {
-        $pesterConfiguration = [PesterConfiguration]::new()
+        $pesterConfiguration = New-PesterConfiguration
         $pesterConfiguration.run.Path = $script:UnitTestsPath
         $pesterConfiguration.Run.PassThru = $true
         $pesterConfiguration.Run.Exit = $false
@@ -323,7 +323,7 @@ Add-BuildTask DevCC {
     Remove-Module -Name Pester -Force -ErrorAction 'SilentlyContinue'# there are instances where some containers have Pester already in the session
     Import-Module -Name Pester -MinimumVersion $script:PesterVersion -ErrorAction 'Stop'
 
-    $pesterConfiguration = [PesterConfiguration]::new()
+    $pesterConfiguration = New-PesterConfiguration
     $pesterConfiguration.run.Path = $script:UnitTestsPath
     $pesterConfiguration.CodeCoverage.Enabled = $true
     $pesterConfiguration.CodeCoverage.Path = "$PSScriptRoot\$ModuleName\*\*.ps1"
@@ -511,7 +511,7 @@ Add-BuildTask InfraTest {
         Remove-Module -Name Pester -Force -ErrorAction 'SilentlyContinue'# there are instances where some containers have Pester already in the session
         Import-Module -Name Pester -MinimumVersion $script:MinPesterVersion -MaximumVersion $script:MaxPesterVersion -ErrorAction 'Stop'
 
-        $pesterConfiguration = [PesterConfiguration]::new()
+        $pesterConfiguration = New-PesterConfiguration
         $pesterConfiguration.run.Path = $script:InfraTestsPath
         $pesterConfiguration.Run.PassThru = $true
         $pesterConfiguration.Run.Exit = $false