+++ src/Catesta/Resources/Module/src/PSModule.Settings.ps1
@@ -1,2 +1,2 @@
-# specify the minumum required major PowerShell version that the build script should validate
+# specify the minimum required major PowerShell version that the build script should validate
 [version]$script:requiredPSVersion = '5.1.0'