+++ src/Catesta/Resources/Module/src/Tests/v4/Unit/Public/Public-Function.Tests.ps1
@@ -18,7 +18,7 @@ InModuleScope '<%=$PLASTER_PARAM_ModuleName%>' {
     #-------------------------------------------------------------------------
     $WarningPreference = "SilentlyContinue"
     #-------------------------------------------------------------------------
-    Describe 'Get-HellowWorld Public Function Tests' -Tag Unit {
+    Describe 'Get-HelloWorld Public Function Tests' -Tag Unit {
         Context 'Error' {
 
             # It 'should ...' {
@@ -39,5 +39,5 @@ InModuleScope '<%=$PLASTER_PARAM_ModuleName%>' {
             } #it
 
         } #context_Success
-    } #describe_Get-HellowWorld
+    } #describe_Get-HelloWorld
 } #inModule